fix(ui): deduplicate utility functions — Fixes EVE-124
Consolidate duplicated utility functions into canonical locations:
- formatDistanceToNow, formatDurationCompact, formatFileSize, formatTokens → lib/formatting.ts
- basename → lib/path-utils.ts (new)
- getWorkflowStatusBadgeVariant, getTaskStatusBadgeVariant, getWorkerStatusBadgeVariant,
  getHealthBadgeVariant, getScheduleExecutionBadgeVariant, getCapabilityStatusBadgeVariant
  → lib/status-utils.ts (new)

Replaced ~30 local function copies across 19 files with imports, removing ~280 lines of duplication.
