+++ apps/ui/src/app/(main)/capabilities/[capabilityId]/page.tsx
@@ -12,17 +12,7 @@ import { ArrowLeft, CircleOff, Wrench, FileText, Code, Link as LinkIcon } from "
 import { CopyButton } from "@/components/ui/copy-button";
 import type { CapabilityStatus, ToolDefinition } from "@/lib/api/types";
 import { getCapabilityIcon } from "@/lib/capability-icons";
-
-function getStatusBadgeVariant(status: CapabilityStatus): "default" | "secondary" | "outline" {
-  switch (status) {
-    case "available":
-      return "default";
-    case "coming_soon":
-      return "secondary";
-    case "deprecated":
-      return "outline";
-  }
-}
+import { getCapabilityStatusBadgeVariant } from "@/lib/status-utils";
 
 function getStatusLabel(status: CapabilityStatus): string {
   switch (status) {
@@ -133,7 +123,7 @@ export default function CapabilityDetailPage({
             <h1 className="text-2xl font-bold flex items-center gap-3">
               {capability.name}
               <CopyButton value={capability.id} />
-              <Badge variant={getStatusBadgeVariant(capability.status)}>
+              <Badge variant={getCapabilityStatusBadgeVariant(capability.status)}>
                 {getStatusLabel(capability.status)}
               </Badge>
             </h1>
@@ -206,7 +196,7 @@ export default function CapabilityDetailPage({
             <CardContent className="space-y-4">
               <div>
                 <p className="text-sm font-medium">Status</p>
-                <Badge variant={getStatusBadgeVariant(capability.status)} className="mt-1">
+                <Badge variant={getCapabilityStatusBadgeVariant(capability.status)} className="mt-1">
                   {getStatusLabel(capability.status)}
                 </Badge>
               </div>
