+++ apps/ui/src/app/(main)/durable/page.tsx
@@ -27,6 +27,7 @@ import {
 import Link from "next/link";
 import { Button } from "@/components/ui/button";
 import { CopyButton } from "@/components/ui/copy-button";
+import { getHealthBadgeVariant, getWorkflowStatusBadgeVariant } from "@/lib/status-utils";
 
 function getHealthStatusColor(status: string) {
   switch (status) {
@@ -38,19 +39,6 @@ function getHealthStatusColor(status: string) {
       return "bg-red-500";
     default:
       return "bg-gray-500";
-  }
-}
-
-function getHealthBadgeVariant(status: string) {
-  switch (status) {
-    case "healthy":
-      return "default" as const;
-    case "degraded":
-      return "secondary" as const;
-    case "unhealthy":
-      return "destructive" as const;
-    default:
-      return "outline" as const;
   }
 }
 
@@ -396,7 +384,7 @@ export default function DurableDashboardPage() {
                           <CopyButton value={workflow.id} />
                         </div>
                       </div>
-                      <Badge variant={getWorkflowStatusVariant(workflow.status)}>
+                      <Badge variant={getWorkflowStatusBadgeVariant(workflow.status)}>
                         {workflow.status}
                       </Badge>
                     </Link>
@@ -456,17 +444,3 @@ function WorkflowStatusIcon({ status }: { status: string }) {
   }
 }
 
-function getWorkflowStatusVariant(status: string) {
-  switch (status) {
-    case "completed":
-      return "default" as const;
-    case "running":
-      return "secondary" as const;
-    case "failed":
-      return "destructive" as const;
-    case "cancelled":
-      return "outline" as const;
-    default:
-      return "outline" as const;
-  }
-}
