+++ apps/ui/src/app/(main)/durable/schedules/[scheduleId]/page.tsx
@@ -69,31 +69,8 @@ import {
 } from "lucide-react";
 import Link from "next/link";
 import { CopyButton } from "@/components/ui/copy-button";
-
-function formatDistanceToNow(date: Date, options?: { addSuffix?: boolean }): string {
-  const now = new Date();
-  const diff = now.getTime() - date.getTime();
-  const seconds = Math.floor(Math.abs(diff) / 1000);
-  const minutes = Math.floor(seconds / 60);
-  const hours = Math.floor(minutes / 60);
-  const days = Math.floor(hours / 24);
-
-  let result = "";
-  if (days > 0) {
-    result = `${days} day${days > 1 ? "s" : ""}`;
-  } else if (hours > 0) {
-    result = `${hours} hour${hours > 1 ? "s" : ""}`;
-  } else if (minutes > 0) {
-    result = `${minutes} minute${minutes > 1 ? "s" : ""}`;
-  } else {
-    result = "less than a minute";
-  }
-
-  if (options?.addSuffix) {
-    return diff > 0 ? `${result} ago` : `in ${result}`;
-  }
-  return result;
-}
+import { formatDistanceToNow } from "@/lib/formatting";
+import { getScheduleExecutionBadgeVariant } from "@/lib/status-utils";
 
 function getStatusIcon(status: ScheduleExecutionStatus) {
   switch (status) {
@@ -111,28 +88,13 @@ function getStatusIcon(status: ScheduleExecutionStatus) {
   }
 }
 
-function getStatusBadgeVariant(status: ScheduleExecutionStatus) {
-  switch (status) {
-    case "completed":
-      return "default" as const;
-    case "running":
-      return "secondary" as const;
-    case "failed":
-      return "destructive" as const;
-    case "skipped":
-    case "pending":
-    default:
-      return "outline" as const;
-  }
-}
-
 function ExecutionRow({ execution }: { execution: ScheduleExecution }) {
   return (
     <TableRow>
       <TableCell>
         <div className="flex items-center gap-2">
           {getStatusIcon(execution.status)}
-          <Badge variant={getStatusBadgeVariant(execution.status)}>{execution.status}</Badge>
+          <Badge variant={getScheduleExecutionBadgeVariant(execution.status)}>{execution.status}</Badge>
         </div>
       </TableCell>
       <TableCell>
