+++ apps/ui/src/components/chat/read-file-tool-call-card.tsx
@@ -11,6 +11,8 @@ import { useMemo, useState } from "react";
 import { Check, ChevronDown, ChevronRight, FileText, Loader2 } from "lucide-react";
 import type { ContentPart, ToolCompletedData } from "@/lib/api/types";
 import { cn } from "@/lib/utils";
+import { basename } from "@/lib/path-utils";
+import { formatFileSize } from "@/lib/formatting";
 import { getFullText, type ToolCallContent } from "./tool-call-utils";
 
 interface ReadFileToolCallCardProps {
@@ -37,12 +39,6 @@ interface ParsedReadFileResult {
   encoding?: string;
   sizeBytes?: number;
   images: ParsedReadFileImage[];
-}
-
-function basename(value: string): string {
-  const clean = value.replace(/\/+$/, "");
-  const parts = clean.split("/");
-  return parts[parts.length - 1] || clean;
 }
 
 function getPathFromArguments(toolCall: ToolCallContent): string | undefined {
@@ -131,13 +127,6 @@ function getContentPreview(content: string | null): string | null {
     : firstNonEmptyLine;
 }
 
-function formatSize(sizeBytes: number | undefined): string | null {
-  if (typeof sizeBytes !== "number" || sizeBytes < 0) return null;
-  if (sizeBytes < 1024) return `${sizeBytes} B`;
-  if (sizeBytes < 1024 * 1024) return `${(sizeBytes / 1024).toFixed(1)} KB`;
-  return `${(sizeBytes / (1024 * 1024)).toFixed(1)} MB`;
-}
-
 export function isReadFileTool(toolName: string): boolean {
   return toolName === "read_file" || toolName === "session_read_file";
 }
@@ -164,7 +153,7 @@ export function ReadFileToolCallCard({ toolCall, toolResult }: ReadFileToolCallC
     getContentPreview(parsed.content) ??
     (hasImages ? `${parsed.images.length} image${parsed.images.length === 1 ? "" : "s"}` : null) ??
     (isBinaryWithoutPreview
-      ? `binary file${formatSize(parsed.sizeBytes) ? ` (${formatSize(parsed.sizeBytes)})` : ""}`
+      ? `binary file${formatFileSize(parsed.sizeBytes) ? ` (${formatFileSize(parsed.sizeBytes)})` : ""}`
       : null);
   const title = parsed.path ? `Read ${basename(parsed.path)}` : "Read file";
   const showDetails = hasImages || isExpanded;
